requirements.txt
streamlit==1.22.0
tensorflow==2.10.0
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
protobuf==3.19.6